requirements.txt
# Core packages
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Image processing
scikit-image>=0.18.0
opencv-python>=4.5.0
pillow>=8.3.0
albumentations>=1.1.0

# Deep learning
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.8.0
monai>=1.0.0  # Medical imaging specific library

# Machine learning and evaluation
scikit-learn>=1.0.0
xgboost>=1.5.0
optuna>=3.0.0  # Hyperparameter optimization
shap>=0.41.0   # Model interpretability

# Data management and tracking
wandb>=0.13.0  # Experiment tracking
h5py>=3.6.0
pydicom>=2.3.0

# Other utilities
tqdm>=4.62.0
click>=8.0.0
jupyterlab>=3.4.0